Write a program to evaluate following investment equation V = P(1+r)n and print the tables which would give the value of V for various combination of the  following values of P, r, and n:  P: 1000, 2000, 3000, ........, 10000  r: 0.10, 0.12, 0.13, ........, 0.20  n: 1, 2, 3, ........, 10  Hint:- P is the principal amount and v is the Value of money at the end of n years. This equation can be recursively written as V = P(1+r) P = V
The C++ program calculates and prints the values of an investment equation for various combinations of principal amount (
�
P), interest rate (
�
r), and time (
�
n). The investment equation 
�
=
�
(
1
+
�
)
�
V=P(1+r) 
n
  represents the future value of an investment based on its principal, interest rate, and time period. The program utilizes nested loops to iterate through predefined ranges of 
�
P, 
�
r, and 
�
n, calculating the corresponding 
�
V for each combination and printing the results in a table format. The output provides insight into how different values of principal, interest rate, and time impact the future value of the investment.